day-3:control flow with if else and condition operation

+++ day-2/tip-calculator.py
@@ -1,5 +1,5 @@
 print("Welcome to the tip calculator.")
-bill=int(input("What was the total bill ? "))
+bill=float(input("What was the total bill ? "))
 print(bill)
 
 tip=int(input("What percentage tip would you like to give ? "))
